Right hand menu implementation
Enabled the metronome to be toggled with the right hand

Modified the Metronome class to use the gamecontroller to be more
accessable to the rest of the scene
Metronome button calls ToggleMetronome() in gamecontroller

added REC button to record  looping
REC button calls ToggleRecLoop() in the gamecontroller.
Added a boolean variable for the REC function to toggle to be more
accessible

Added Play button to right hand menu
Play button calls TogglePlayLoop() in the gamecontroller
added boolean for the same reason as the other actions.