Drumstick meshes were added
I added the drumstick meshes. Now for some reason the drum hitboxes are
messing up.

I think it is because the drums were missing their scripts on mine
(similar to how the game controller is always missing its scripts) and
so they could not call the on-trigger enter function. maybe...